Authenticate in all methods + fixes in users/events/notifications

+++ Time_planner_api/Time_planner_api/Controllers/EventsController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -12,9 +13,8 @@ namespace Time_planner_api.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class EventsController : ControllerBase
+    public class EventsController : BaseController
     {
-
         private readonly DatabaseContext _context;
 
         public EventsController(DatabaseContext context)
@@ -22,51 +22,150 @@ namespace Time_planner_api.Controllers
             _context = context;
         }
 
-
-        // GET: api/Events/104416411457610
-        [HttpGet("{userId}")]
-        public async Task<ActionResult<IEnumerable<Event>>> GetEvents(string userId)
+        // GET: api/Events/user?id=a
+        [HttpGet]
+        [Authorize]
+        [Route("user")]
+        public async Task<ActionResult<IEnumerable<Event>>> GetEvents(string id)
         {
-            return await _context.Events.Where(e => e.OwnerId == userId).ToListAsync();          
+            var allEvents = await _context.Events.Where(e => e.OwnerId == id).ToListAsync();
+            if (GetUserId() == id)
+            {
+                return allEvents;
+            }
+
+            var filteredEvents = new List<Event>();
+            foreach(var ev in allEvents)
+            {
+                if (ev.IsPublic)
+                {
+                    filteredEvents.Add(new Event()
+                    {
+                        Id = ev.Id,
+                        Title = ev.Title,
+                        StartDate = ev.StartDate,
+                        EndDate = ev.EndDate
+                    });
+                }
+                else
+                {
+                    filteredEvents.Add(new Event()
+                    {
+                        Id = ev.Id,
+                        Title = "",
+                        StartDate = ev.StartDate,
+                        EndDate = ev.EndDate
+                    });
+                }
+            }
+
+            return filteredEvents;
         }
 
-
-        // GET: api/Events/104416411457610/2
-        [HttpGet("{userId}/{id}")]
-        public async Task<ActionResult<Event>> GetEvent(string userId, int id)
+        // GET: api/Events/event?id=a
+        [HttpGet]
+        [Authorize]
+        [Route("event")]
+        public async Task<ActionResult<Event>> GetEvent(int id)
         {
             var ourEvent = await _context.Events.FindAsync(id);
+            var userId = GetUserId();
 
             if (ourEvent == null)
             {
                 return NotFound();
             }
 
-            return ourEvent;
+            if (ourEvent.IsPublic || userId == ourEvent.OwnerId)
+            {
+                return ourEvent;
+            }
+            else
+            {
+                var invitedFriends = await _context.UsersEvents.Where(e => e.EventId == id && e.Status != 2 /* not rejected */).Select(ee => ee.UserId).ToListAsync();
+                if (invitedFriends.Contains(userId))
+                {
+                    return new Event()
+                    {
+                        Id = ourEvent.Id,
+                        Title = ourEvent.Title,
+                        StartDate = ourEvent.StartDate,
+                        EndDate = ourEvent.EndDate
+                    };
+                }
+                return new Event()
+                {
+                    Id = ourEvent.Id,
+                    Title = "",
+                    StartDate = ourEvent.StartDate,
+                    EndDate = ourEvent.EndDate
+                };
+            }
         }
 
         // POST: api/Events
         [HttpPost]
-        public async Task<ActionResult<Event>> PostEvent(Event ourEvent)
+        [Authorize]
+        public async Task<ActionResult<Event>> PostEvent(EventWithFriends ourEvent)
         {
-            await _context.Events.AddAsync( new Event() {
-                StartDate = ourEvent.StartDate.ToLocalTime(),
-                Title = ourEvent.Title, 
-                EndDate= ourEvent.EndDate.ToLocalTime(),
-                City = ourEvent.City,
-                StreetAddress = ourEvent.StreetAddress,
-                IsPublic = ourEvent.IsPublic,
-                OwnerId = ourEvent.OwnerId,
-                Description = ourEvent.Description
-            }); 
+            if (ourEvent.Event.OwnerId != GetUserId())
+            {
+                return Unauthorized();
+            }
+
+            var newEvent = new Event()
+            {
+                StartDate = ourEvent.Event.StartDate.ToLocalTime(),
+                Title = ourEvent.Event.Title,
+                EndDate = ourEvent.Event.EndDate.ToLocalTime(),
+                City = ourEvent.Event.City,
+                StreetAddress = ourEvent.Event.StreetAddress,
+                IsPublic = ourEvent.Event.IsPublic,
+                OwnerId = ourEvent.Event.OwnerId,
+                Description = ourEvent.Event.Description
+            };
+
+            await _context.Events.AddAsync(newEvent); 
 
             try
             {
+                // save event
+                await _context.SaveChangesAsync();
+                var eventId = newEvent.Id;
+                // add usersEvents for owner
+                await _context.UsersEvents.AddAsync(new UsersEvents()
+                {
+                    EventId = eventId,
+                    UserId = ourEvent.Event.OwnerId,
+                    Status = 3 // owner
+                });
+                if (ourEvent.FriendIds != null)
+                {
+                    foreach (var friendId in ourEvent.FriendIds)
+                    {
+                        // add usersEvents for friends
+                        await _context.UsersEvents.AddAsync(new UsersEvents()
+                        {
+                            EventId = eventId,
+                            UserId = friendId,
+                            Status = 0 // unknown
+                        });
+                        // add notifications for friends
+                        await _context.Notifications.AddAsync(new Notification()
+                        {
+                            IsDismissed = false,
+                            MessageType = 0, // invited
+                            SenderId = ourEvent.Event.OwnerId,
+                            ReceiverId = friendId,
+                            EventId = eventId
+                        });
+                    }
+                }
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateException)
             {
-                if (EventExists(ourEvent.Id))
+                if (EventExists(ourEvent.Event.Id))
                 {
                     return Conflict();
                 }
@@ -76,13 +175,20 @@ namespace Time_planner_api.Controllers
                 }
             }
 
-            return CreatedAtAction("GetEvent", new {userId = ourEvent.OwnerId, id = ourEvent.Id }, ourEvent);
+            return CreatedAtAction("GetEvent", new {userId = ourEvent.Event.OwnerId, id = ourEvent.Event.Id }, ourEvent);
         }
 
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<ActionResult> PutEvent([FromRoute]int id, Event oldEvent)
         {
             Event newEvent = _context.Events.Where(e => e.Id == id).Single<Event>();
+
+            if (GetUserId() != oldEvent.OwnerId || oldEvent.OwnerId != newEvent.OwnerId)
+            {
+                return Unauthorized();
+            }
+
             newEvent.Id = oldEvent.Id;
             newEvent.StartDate = oldEvent.StartDate.ToLocalTime();
             newEvent.Title = oldEvent.Title;
@@ -116,6 +222,7 @@ namespace Time_planner_api.Controllers
 
         // DELETE: api/Events/5
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<ActionResult<Event>> DeleteEvent([FromRoute] int id)
         {
             var ourEvent = await _context.Events.FindAsync(id);
@@ -124,7 +231,14 @@ namespace Time_planner_api.Controllers
                 return NotFound();
             }
 
+            if (GetUserId() != ourEvent.OwnerId)
+            {
+                return Unauthorized();
+            }
+
             _context.Events.Remove(ourEvent);
+            var notifications = _context.Notifications.Where(n => n.EventId == id);
+            _context.Notifications.RemoveRange(notifications);
             await _context.SaveChangesAsync();
 
             return ourEvent;
@@ -137,5 +251,3 @@ namespace Time_planner_api.Controllers
       
     }
 }
-
-
